add delay information

+++ include/ublksrv_delay.h
@@ -1,5 +1,8 @@
 #ifndef UBLKSRV_DELAY_INC_H
 #define UBLKSRV_DELAY_INC_H
-extern int ublksrv_get_current_tick();
+
+
+extern uint64_t ublksrv_get_current_tick();
 extern int ublksrv_delay_module(int ublk_op);
 #endif
+
